pyproject.toml
[project]
name = "valorant-tracker-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "psycopg2-binary>=2.9.7",
    "sqlalchemy>=2.0.36",
    "pydantic>=2.7.2",
    "python-dotenv>=1.0.0",
    "discord.py>=2.3.2",
    "apscheduler>=3.10.4",
]

[tool.setuptools.packages.find]
where = ["src"]
